refactor(laguna): move LagunaS21Renderer to its own module
Review feedback: S-2.1 didn't belong in laguna_xs2.py. Pure code motion to
renderers/laguna_s21.py, following the qwen35.py -> qwen36.py precedent for a
cross-module renderer subclass. Updated the _LAZY_RENDERERS entry, the
_populate_registry import, the config docstring cross-ref, and the test import.

laguna_xs2.py still imports LagunaS21RendererConfig for two config-union
annotations — the base __init__s must accept an S-2.1 config forwarded up
through super().
